ds3: use "retry" from uni_hid_device
Instead of using the custom DS3 retry logic, it call the
uni_hid_device_queue_* function. That function does the retries in a
more efficient way since it waits until the controllers tells that it is
ready to send.

Additionally it fixes a bug in circular buffer. Don't know how it was
working before \O/