unijoysticle2plus: cp2102, add extra 7405 and more
cp2102:
JLCPCB has the cp2102 as basic part. Cost wise, it might be a tiny bit
more expensive, but the footprint is smaller.
The downsize is that this is the cp2102, and not the recommended
cp2102N.

extra 74HC05:
Added another 74HC05 to connect POT-X, and POT-Y of Joy2. This could be
useful for Amiga computers, where some games (?) might use a 2nd and 3rd
buttons (?).

Add missing caps to ESP32, fixed some values, exposed GPIOs 36 and 39...