unijoysticle 2+: use better USB micro B connector
The one used in Rev F works Ok when the Fab house solder it correctly.
But it might be possible that it might not work. At least I found 1 / 20
that was not working due to a bad solder/alignment (?) in one of the
boards.

So, I'm switching to one that has some holes so the alignment should not
fail. It might required some extra manual solder... but we shall see.

Also, I slightly improved the routing traces between the USB D- / D+
with the CH340 chip.

Also, I'm using 1.5K resitor for Green LED as well.
The previous LED that was used in Rev F is no longer available and the
2nd batch came with a more potent one. So I need to increase the
resistor.

Besides those two little changes, everything is the same.
Revistion G and F are functionality-wise the same.