unijoysticle2+: kicad 6 migration
Migrates Unijoysticle 2+ Kicad 5 files to Kicad 6

In the process, it bumps version to Rev. H.
The schematic is the "same": just minor cosmetic changes were done only
to make ERC happy.

Similar on the PCB: it is the same, but NET-R5-PAD2 and NET-R6-PAD2 were
re-routed again since "Tools -> Clean up vias and tracks" removed it.

Besides that, the shcematic and layout are the same...