uni2 a500: revert CH340 from 5V
Autoreset is driven by DTS/RTS.
But if DTS/RTS are not used (requires that CH340 is powered) then
autoreset doesn't work.
So, although we shouldn't power CH340 when it is powered from the Amiga,
since it consumes current, we need it for autoreset.

Besides, we don't gain much by not powering CH340. Only 0.02 amps
according to my measures.